semver: forward to x/mod/semver (#210)
This is now the canonical location of the semver package.
